backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.7.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
redis==5.2.1
celery==5.4.0
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
stripe==11.4.0
python-multipart==0.0.19
httpx==0.28.1
